requirements.txt
streamlit
opencv-python
numpy
Pillow
pywavelets
scikit-image
boto3
pycryptodome